requirements.txt
flask==2.3.3
ultralytics==8.0.196
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
scikit-learn==1.3.0
tqdm==4.66.1
pyyaml==6.0.1
requests==2.31.0
selenium==4.11.2
beautifulsoup4==4.12.2
transformers==4.33.2
clip-by-openai==1.2
psutil==5.9.5
gunicorn==21.2.0